refactor(sema): move builtin index impl
